backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
firebase-admin==6.5.0
pydantic==2.6.4
python-multipart==0.0.9
gunicorn==21.2.0
python-dotenv==1.0.1
